+++ src/main/java/com/tinyhuman/tinyhumanapi/album/infrastructure/AlbumRepositoryImpl.java
@@ -3,7 +3,7 @@ package com.tinyhuman.tinyhumanapi.album.infrastructure;
 
 import com.tinyhuman.tinyhumanapi.album.domain.Album;
 import com.tinyhuman.tinyhumanapi.album.service.port.AlbumRepository;
-import com.tinyhuman.tinyhumanapi.common.domain.exception.ResourceNotFoundException;
+import com.tinyhuman.tinyhumanapi.common.exception.ResourceNotFoundException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Repository;
