+++ src/main/java/com/tinyhuman/tinyhumanapi/album/service/AlbumServiceImpl.java
@@ -6,8 +6,8 @@ import com.tinyhuman.tinyhumanapi.album.controller.port.AlbumService;
 import com.tinyhuman.tinyhumanapi.album.domain.Album;
 import com.tinyhuman.tinyhumanapi.album.service.port.AlbumRepository;
 import com.tinyhuman.tinyhumanapi.auth.controller.port.AuthService;
-import com.tinyhuman.tinyhumanapi.common.domain.exception.ResourceNotFoundException;
-import com.tinyhuman.tinyhumanapi.common.domain.exception.UnauthorizedAccessException;
+import com.tinyhuman.tinyhumanapi.common.exception.ResourceNotFoundException;
+import com.tinyhuman.tinyhumanapi.common.exception.UnauthorizedAccessException;
 import com.tinyhuman.tinyhumanapi.common.enums.ContentType;
 import com.tinyhuman.tinyhumanapi.integration.service.ImageService;
 import com.tinyhuman.tinyhumanapi.integration.util.ImageUtil;
