+++ src/main/java/com/tinyhuman/tinyhumanapi/auth/service/AuthServiceImpl.java
@@ -6,7 +6,7 @@ import com.tinyhuman.tinyhumanapi.auth.domain.LoginRequest;
 import com.tinyhuman.tinyhumanapi.auth.domain.RefreshToken;
 import com.tinyhuman.tinyhumanapi.auth.domain.TokenResponse;
 import com.tinyhuman.tinyhumanapi.auth.service.port.RefreshTokenRepository;
-import com.tinyhuman.tinyhumanapi.common.domain.exception.ResourceNotFoundException;
+import com.tinyhuman.tinyhumanapi.common.exception.ResourceNotFoundException;
 import com.tinyhuman.tinyhumanapi.user.domain.User;
 import com.tinyhuman.tinyhumanapi.user.service.port.UserRepository;
 import lombok.RequiredArgsConstructor;
