+++ src/main/java/com/tinyhuman/tinyhumanapi/common/exception/UnauthorizedAccessException.java
@@ -1,4 +1,4 @@
-package com.tinyhuman.tinyhumanapi.common.domain.exception;
+package com.tinyhuman.tinyhumanapi.common.exception;
 
 public class UnauthorizedAccessException extends RuntimeException {
