+++ src/main/java/com/tinyhuman/tinyhumanapi/user/service/UserServiceImpl.java
@@ -1,6 +1,6 @@
 package com.tinyhuman.tinyhumanapi.user.service;
 
-import com.tinyhuman.tinyhumanapi.common.domain.exception.ResourceNotFoundException;
+import com.tinyhuman.tinyhumanapi.common.exception.ResourceNotFoundException;
 import com.tinyhuman.tinyhumanapi.user.controller.port.UserService;
 import com.tinyhuman.tinyhumanapi.user.domain.User;
 import com.tinyhuman.tinyhumanapi.user.domain.UserCreate;
