+++ src/test/java/com/tinyhuman/tinyhumanapi/album/service/AlbumServiceImplTest.java
@@ -9,7 +9,7 @@ import com.tinyhuman.tinyhumanapi.baby.domain.Baby;
 import com.tinyhuman.tinyhumanapi.baby.enums.Gender;
 import com.tinyhuman.tinyhumanapi.baby.mock.FakeImageService;
 import com.tinyhuman.tinyhumanapi.baby.mock.FakeMultipartFile;
-import com.tinyhuman.tinyhumanapi.common.domain.exception.UnauthorizedAccessException;
+import com.tinyhuman.tinyhumanapi.common.exception.UnauthorizedAccessException;
 import com.tinyhuman.tinyhumanapi.common.enums.ContentType;
 import com.tinyhuman.tinyhumanapi.user.domain.User;
 import com.tinyhuman.tinyhumanapi.user.domain.UserBabyRelation;
