+++ src/test/java/com/tinyhuman/tinyhumanapi/diary/service/DiaryServiceImplTest.java
@@ -6,8 +6,8 @@ import com.tinyhuman.tinyhumanapi.baby.enums.Gender;
 import com.tinyhuman.tinyhumanapi.baby.mock.FakeBabyRepository;
 import com.tinyhuman.tinyhumanapi.baby.mock.FakeImageService;
 import com.tinyhuman.tinyhumanapi.baby.mock.FakeMultipartFile;
-import com.tinyhuman.tinyhumanapi.common.domain.exception.ResourceNotFoundException;
-import com.tinyhuman.tinyhumanapi.common.domain.exception.UnauthorizedAccessException;
+import com.tinyhuman.tinyhumanapi.common.exception.ResourceNotFoundException;
+import com.tinyhuman.tinyhumanapi.common.exception.UnauthorizedAccessException;
 import com.tinyhuman.tinyhumanapi.diary.domain.Diary;
 import com.tinyhuman.tinyhumanapi.diary.domain.DiaryCreate;
 import com.tinyhuman.tinyhumanapi.diary.domain.DiaryResponse;
